+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [0.5.0] - 2022-12-16
+
 ### Added
 
 - Made built in escaping rules less aggressive for serialization; text values
@@ -56,5 +58,6 @@
 Initial public release.
 
 <!-- next-url -->
+[Unreleased]: https://github.com/faassen/xot/compare/v0.5.0...HEAD
 
-[unreleased]: https://github.com/faassen/xot/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/faassen/xot/compare/v0.4.0...v0.5.0